small fix int to long

+++ src/main/kotlin/com/akopyan/exelparser/domain/database/EmployeesRepo.kt
@@ -3,6 +3,6 @@ package com.akopyan.exelparser.domain.database
 import org.springframework.data.repository.CrudRepository
 
 
-interface EmployeesRepo : CrudRepository<Employees, Int> {
+interface EmployeesRepo : CrudRepository<Employees, Long> {
     fun findAllByToken(token: String): List<Employees>
 }
